+++ src/frontend/src/tests/mocks/exchanges.mock.ts
@@ -1,6 +1,5 @@
 import type { CoingeckoSimpleTokenPrice } from '$lib/types/coingecko';
 import type { ExchangesData } from '$lib/types/exchange';
-import type { KongSwapToken, KongSwapTokenMetrics } from '$lib/types/kongswap';
 import { mockTokens } from './tokens.mock';
 
 export const mockOneUsd = 1;
@@ -9,51 +8,6 @@ export const mockExchanges: ExchangesData = mockTokens.reduce<ExchangesData>((ac
 	acc[token.id] = { usd: mockOneUsd };
 	return acc;
 }, {});
-
-export const createMockKongSwapToken = ({
-	token = {},
-	metrics = {}
-}: {
-	canisterId?: string;
-	token?: Partial<KongSwapToken['token']>;
-	metrics?: Partial<KongSwapTokenMetrics>;
-}): KongSwapToken => {
-	const tokenDefaults = {
-		token_id: 1,
-		name: 'Mock Token',
-		symbol: 'MOCK',
-		canister_id: 'aaaaa-aa',
-		address: '0xABC123',
-		decimals: 8,
-		fee: 0.0001,
-		fee_fixed: '10',
-		has_custom_logo: false,
-		icrc1: true,
-		icrc2: true,
-		icrc3: false,
-		is_removed: false,
-		logo_url: null,
-		logo_updated_at: '2024-01-01T00:00:00.000Z',
-		token_type: 'Ic'
-	};
-
-	const metricsDefaults: KongSwapTokenMetrics = {
-		token_id: 1,
-		price: '1.00',
-		market_cap: '1000000',
-		volume_24h: '50000',
-		price_change_24h: '2.5',
-		updated_at: '2024-01-01T00:00:00.000Z',
-		total_supply: '100000000',
-		tvl: '250000',
-		previous_price: '0.95'
-	};
-
-	return {
-		token: { ...tokenDefaults, ...token },
-		metrics: { ...metricsDefaults, ...metrics }
-	};
-};
 
 export const createMockCoingeckoTokenPrice = (
 	overrides: Partial<CoingeckoSimpleTokenPrice> = {}
